requirements-test.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
black==19.3b0
cattrs==0.9.0
Click==7.0
dataclasses==0.6
decorator==4.3.2
django-stubs==0.10.2
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
more-itertools==6.0.0
mypy==0.670
mypy-extensions==0.4.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.9.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.3.1
pytest==4.3.1
pytest-django==3.4.8
six==1.12.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.3.1
typing-extensions==3.7.2
wcwidth==0.1.7
